requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
discord.py==1.7.3
idna==3.3
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.1
regex==2021.10.23
tomli==1.2.2
typing-extensions==3.10.0.2
yarl==1.7.0
